Got basic dummy edges to work, fix coord types to i32

+++ yew-flow/src/components/node/models.rs
@@ -15,8 +15,8 @@ pub struct NodeOutput {
 pub struct Node {
     pub id: usize,
     pub title: String,
-    pub x: u64,
-    pub y: u64,
+    pub x: i32,
+    pub y: i32,
     pub color: Hsl,
     pub is_active: bool,
     pub inputs: Vec<NodeInput>,
@@ -28,13 +28,13 @@ pub struct Node {
 pub struct Edge {
     pub id: usize,
     /// From (x1, y1) coordinate
-    pub x1: u64,
+    pub x1: i32,
     /// From (x1, y1) coordinate
-    pub y1: u64,
+    pub y1: i32,
     /// To (x2, y2) coordinate
-    pub x2: u64,
+    pub x2: i32,
     /// To (x2, y2) coordinate
-    pub y2: u64,
+    pub y2: i32,
     pub color: Hsl,
     pub is_active: bool,
 }
